Calculates the proportion of individuals that leave a size class

The proportion of individuals that leave each size class. These proportion are scaled to the fastest growing species/size class combination. This ensures that no species can grow though more than one class interval in any time step. This is based on the time (yrs) for an average fish to grow from the lower limit of a size class to the upper limit of a size class (Hilborn & Walters p428, eqn 13.7.2).

calc_phi(nSize, nSpecies, uBound, lBound, parameterValues)

Arguments

nSize

Number of size class intervals species can grow through

nSpecies

Number of species in the model

uBound

Upper bound of each size class interval

lBound

Lower bound of each size class interval

parameterValues

A Matrix of species specific parameters. See rochet_GB_parameterValues

Value

A list is returned

probGrowOut

A matrix of size nSize x nSpecies of proportions. A cell represent the proportion of individuals growing into the following cell. Note: probGrowOut = 0 for size classes in which a species does not reach (due to its L_inf). probGrowOut = 0 for largest size class for each species

phiMin

Model timestep (years). Shhortest time taken by any species to grow into next size class. probGrowOut is normalized by this quantity)

References

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486