Calculates recruits and spawning stock biomass (SSB)

The number of recruits are estimated as a function of SSB using the Ricker stock-recruitment relationship. The SSB is estimated as a function of the number of mature individuals.

calc_recruits(N, maturity, wgt, recAlpha, recBeta)

Arguments

N

A matrix (nSize x nSpecies) of abundance (number of individuals)

maturity

A matrix (nSize x nSpecies) of the proportion of individuals that are considered mature (calc_maturity)

wgt

A matrix (nSize x nSpecies) of an individuals weight at the mid point of each size class (Units: grams). See calc_ration

recAlpha

A vector (length nSpecies). Scale parameter of the Ricker S-R relationship.

recBeta

A vector (length nSpecies). Shape parameter of the Ricker S-R relationship.

Value

A list is returned

recruits

A vector (length nSpecies). Number of recruits (millions).

SSB

A vector (length nSpecies). SSB (1000's tonnes).

References

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486