Key_Run (Length-based Multispecies Analysis by numerical Simulation)

Runs the LeMANS model with the bundled data. Bundled data conist of the 22 species used in the Rochet et al (2011) paper to represent Georges Bank. See rochet_GB_species. The code was adapted directly from MATLAB code used in the Hall et al (2006) and Rochet et al (2011) papers.

key_run(Ffull, nYrs, modelSetup, parameterValues, initialValues, foodweb,
  species)

Arguments

Ffull
Fishing mortaliy rate for a fully recruited fish

nYrs
Number of years to simulate.

modelSetup
list of parameters required for model (See rochet_GB_modelSetup)

parameterValues
matrix of species specific parameters (See rochet_GB_parameterValues)

initialValues
matrix of initial abundance estimates (See rochet_GB_initialValues)

foodweb
predator prey food web (See rochet_GB_foodweb)

species
matrix of species names and guild membership (See rochet_GB_species)

Value

A list containing:

N

3D array of abundance (numbers of animals). nsizeClass x nSpecies x nTimeSteps

M1

Matrix of M1 mortality ("natural"). nsizeClass x nSpecies

M2

3D array M2 mortality (predation). nsizeClass x nSpecies x nTimeSteps

F

3D array M2 mortality (predation). nsizeClass x nSpecies x nTimeSteps

catch

3D array of catch (numbers of animals). nsizeClass x nSpecies x nTimeSteps

SSB

Matrix of spawning stock biomass (SSB). nSpecies x nYears

recruits

Matrix of recruits (numbers of animals). nSpecies x nYears

suitability

3D array of prey suitability by predator. nsizeClass x nSpecies x (num Pred.size class combinations). See calc_sizepref_suitability

sizePreference

3D array of prey preference by predator. nsizeClass x nSpecies x (num Pred.size class combinations). See calc_sizepref_suitability

ration

Matrix. Amount consumed to account for growth. See calc_phi

growthEfficiency

Matrix of growth efficiencies. See calc_phi

growthProportions

Matrix of proportions. Proportion of individuals that leave each size class in each time step. See calc_phi

maturityProportions

Matrix of maturity proportions. See calc_maturity

modelTimeStep

scalar representing the fraction of a year each time step represents. See calc_phi

Using the Rochet et al data set

The number of size classes and the width of the size class was decided upon a priori. This decision was based on the maximum L_inf among all species. max(Linf) = 148 cm The rochet_GB_initialValues should be set up to represent the number of size classes and width.

The unit of output for Rochet et al data:

catch, N: number of individuals

M1, M2, eF: rates

recruits and SSB: were scaled to common recruit and spawning stock size units (individuals × 10^6 and tonnes × 10^3, respectively)

See also

Examples

# NOT RUN {
#'# runs the model with bundled data from Rochet et al (2011).
output <- key_run(Ffull=.4,nYrs=50,rochet_GB_modelSetup,rochet_GB_parameterValues,rochet_GB_initialValues,rochet_GB_foodweb,rochet_GB_species)
# }