List of species independent parameters

A list of parameter values that are not species specific that are required for the model to run. All values are Scalar values.

rochet_GB_modelSetup

Format

A list of species independent parameters:

otherFood

Amount of other food available not accounted for in the model (g)

predationFlag

A flag denoting if predation is on 1 or off 0 in model

Falpha

Steepness of selectivity curve

FL50

Length at 50% selectivity

alphaInt

Scaling for Ricker alpha wrt L_inf

alphaExp

Scaling for Ricker alpha wrt L_inf

betaInt

Scaling for Ricker beta wrt Smax

betaExp

Scaling for Ricker beta wrt Smax

SmaxScale

Catchability scaling for Smax estimate from survery

forageFishAlpha

Trial for forage fish

alphaM1

beta pdf parameter to model M1

betaM1

beta pdf parameter to model M1

cM1

Scaling of pdf to model M1

spMu

Mean of size preference function

spSigma

Standard deviation of size preference function

Source

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63: 1344-1359

Rochet et al. (2011). Does selective fishing conserve community biodiversiy? Predictions from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68: 469-486

See also