Parameter values used in the model

Species specific parameter values

rochet_GB_parameterValues

Format

A data frame of size nSpecies x 8. The columns are defined as:

k

Von Bertalanfy growth parameter (year^-1)

Linf

Maximum length species can grow (cm)

Lmat

Length at maturity (cm)

kappa

Curvature parameter for the maturity ogive

w_a

Parameter of lenth-weight conversion

w_b

Parameter of lenth-weight conversion

Smax

Maximum observed spawning stock biomass (kg)

IsFished

Flag indicating if a species is fished or not

The list of species can be found in rochet_GB_species and in the references below.

Source

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63: 1344-1359

Rochet et al. (2011). Does selective fishing conserve community biodiversiy? Predictions from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68: 469-486

See also