Calculate fishing Mortality

Calculates the fishing mortality for each species in each size class using a logistic selectivity curve.

calc_F(nSize, nSpecies, mBound, lBound, Ffull, Falpha, FL50, scLinf,
  scLinfMat, phiMin, parameterValues)

Arguments

nSize

Number of size class intervals species can grow through

nSpecies

Number of species in the model

mBound

Vector. Mid point of each size class interval

lBound

Vector. Lower bound of each size class interval

Ffull

Scalar. Fishing mortaliy rate for a fully recruited fish

Falpha

Scalar. Number of years to simulate.

FL50

Scalar. The length at which 50 % selection by the fishery occurs.

scLinf

Vector. The size class at which each species reaches L_inf (maximum length)

scLinfMat

Binary matrix indicating which size classes each species occupies

phiMin

Scalar. Model time step (years). See calc_phi

parameterValues

Matrix (nSize x nSpecies). See rochet_GB_parameterValues

Value

A matrix is returned

F

A matrix (size nSize x nSpecies) of F (fishing mortality) values. Note: F = 0 for size classes in which species do not grow (based on von Bertalanfy growth curve)

References

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486