Plots output from the key_run script

Plots catch, N, M1, M2, SSB, recruits, suitability, growth efficiency, ration, maturity, and growth output from the model run. See key_run for details regarding obtaining output. This is a rudimentary tool used for troubleshooting the model. The user is expected to develop their own set of plotting tools for publication purposes

plot_key_run(dataSet, ylabel = "change this label", is.aggregated = F,
  speciesNames, scales = "fixed", speciesSuitability = 1,
  predOrPrey = "Prey")

Arguments

dataSet

The data set to be plotted. See key_run

ylabel

The yaxis label for the figure. Match this with the data.

is.aggregated

A logical value. Determines if aggregation is required over sizeClass. Only valid for N, M2 and catch.

speciesNames

A data frame in the format of rochet_GB_species

scales

Text indicating whether to display yaxis range fixed for all species ("fixed") or not ("free").

speciesSuitability

Either species number or species name. Only required if passing suitability or size preference data. Default = 1.

predOrPrey

character string indicating whether to plot suitabiliy for species as the prey or predator. Values "prey" or "pred". Default = "Prey"

Notes

Not all arguments are required for plotting. speciesSuitability and predOrPrey are only required for plotting suitabilities (and size preferences). If you are plotting any other output data then you can ignore these arguments. In some cases is.aggregated is irrelevant. A warning will let you know that this is so.

The suitability plots (similar for size preference plots) are interpreted in the following way: If predOrPrey = "prey" and speciesSuitability = "Atlantic cod" then each facet represents a predator of cod. Each line drawn represents a size class of the predator and displays its suitability for each size class (x axis) of Cod (prey)

The units of the output variable are explained in key_run.

'@examples